Migrate admin functions into app package
